Modifications in the backtrackstep
I changed it so it always backtracks, also the evolution of the area is now 1/100000 and i put it in a variable so we can easily modify the number.
I could try different relaxation times but i think its not the time to do so.
I also changed the initial condition to one with less vertices, still when evolved it  goes to more vertices due to the increase in area for nu=0.1