New simulation batch
I will do the bead wrappigng simulations now changing KB
The idea will be to compare how much the shape of the membrane changes with this.
Also post processing got updates so it handles better when the simulations are still running
oN MAIN CLUSTER BEAD we also made sure that the output data file is constantly opened so i dont lose data if the simulation runs out of time.