Major work done
New!  dealing with multiple  beads!
The beads can interact with each other and with the membrane.
For interaction between bead we only have Harmonic potential
The bead can have 3 states, default, manual and froze
Default is moving according to dynamics, manual is with a constant velocity and frozen doesnt move (but you can calculate its force)

New : Protocol for pulling

Now for pulling a membrane we use two beads, one inside the membrane and one outside, the one inside interacts with the one inside with a harmonic interaction and does not interact with the membrane, the one inside feels only repulsion to the membrane and the harmonic interaction.

Currently with outer bead fixed in space we can get a tube

New: Remesher working?!
For proper functionality of the remesher we added a connectivity constraint to make edge collapses more safe,.
This way of cancelling the collapsing is very used and until now we can run simulations that before crashed (!!!)

NEW : Force fields

I have been adding features to implement a force field that affects the membrane, in this case there is no backtracking because we cannot associate a potential to the force field. The implementation is still not ready because we are not certain the volume constraint is being fulfilled.